requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
gunicorn==23.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.5
packaging==24.1
sqlparse==0.5.1
tzdata==2024.2
